Moved files to /src
The package structure should have all library code files inside a `/src` directory and the `<library_name>.dart` should only contain `export`s.
