+++ rest_api/lib/src/rest_api_errors.dart
