+++ rest_api/lib/src/rest_api_logger.dart
