+++ rest_api/lib/src/rest_header_provider.dart
