+++ rest_api/lib/src/rest_response.dart
